requirements.txt
Django==1.8
django-tastypie==0.13.3